Getting rid of more -Wconversion warnings.
git-svn-id: https://svn.r-project.org/R/trunk@55540 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/CommandLineArgs.c
@@ -245,7 +245,7 @@ R_common_command_line(int *pac, char **argv, Rstart Rp)
 
 		else if (lval > 500000)
 		    R_ShowMessage(_("WARNING: '--max-ppsize' value is too large: ignored\n"));
-		else Rp->ppsize = lval;
+		else Rp->ppsize = (size_t) lval;
 	    }
 	    else { /* unknown -option */
 		argv[newac++] = *av;
